+++ examples/start.js
@@ -21,7 +21,7 @@ console.log(v.validate({ id: 5, name: "Al", status: true }, schema));
 				2
 			],
 			field: 'name',
-			message: 'The \'name\' field length must be larger than or equal to 3 characters long!'
+			message: 'The \'name\' field length must be greater than or equal to 3 characters long!'
 		}
 	]
 */