+++ test/messages.spec.js
@@ -38,6 +38,9 @@ describe("Test Messages", () => {
 		expect(msg.dateMax).toBeDefined();
 		expect(msg.forbidden).toBeDefined();
 		expect(msg.email).toBeDefined();
+		expect(msg.url).toBeDefined();
+		expect(msg.uuid).toBeDefined();
+		expect(msg.uuidVersion).toBeDefined();
 
 	});